ajout déplacement de bovins

+++ src/projet2025/Evaluation.java
@@ -78,7 +78,7 @@ public class Evaluation {
                 duplicateRate = 0;
                 return;
             }
-            duplicateRate = (double) duplicateDetections / totalCowsDetected + duplicateDetections;
+            duplicateRate = (double)duplicateDetections / ((double)totalCowsDetected + (double)duplicateDetections);
         }
